refactor(tui): colocate Screen::GistDetail into tui/screens/detail.rs
Seventh screen migration for #287 (Phase 2). Moved into the new
tui/screens/detail.rs: handle_key_detail, detail_guard (the #288 shared
guard, GistDetail-exclusive), fork_intent, preview_detail_file,
park_gist_detail_screen (all GistDetail-exclusive helpers), build_gist_detail_vm
(+ build_comments_pane_vm, comments_title_text), render_gist_detail_vm
(+ render_detail_header_vm, render_gist_file_list_vm, render_gist_comments_vm,
detail_focus_tab, detail_focus_tabs_line), detail_palette_items.

Shared-helper notes: group_by_id stays in mod.rs (general-purpose);
render_centered_modal_input/file_list_scroll/file_rows stay in render.rs
(shared with Confirm's render_compact_gist_bg_vm; file_rows bumped
private -> pub(super)); star_toggle_intent (already pub(crate) from the
Gists migration) reused here too. HelpTopic::for_screen's GistDetail arm
is a plain variant reference with no logic to move, so it stays as-is
(matches the Pins/Gists/Revisions arms, not the Config/Help HELP_TOPIC
const pattern).

Pure move — bodies and existing tests unchanged.

Part of #287
