+++ src/tui/render.rs
@@ -44,7 +44,7 @@ fn render_screen_vm(
             super::screens::gists::render_gists_vm(frame, state, gists, chrome, layout)
         }
         super::ScreenVm::GistDetail(detail) => {
-            render_gist_detail_vm(frame, state, detail, chrome, layout)
+            super::screens::detail::render_gist_detail_vm(frame, state, detail, chrome, layout)
         }
         super::ScreenVm::Revisions(revs) => render_revisions_vm(frame, state, revs, chrome, layout),
         super::ScreenVm::Config(config) => {
@@ -414,7 +414,7 @@ pub(super) fn file_list_scroll(cursor: usize, visible_rows: usize, count: usize)
 /// compaction-confirm background). The first nine files are numbered to match the 1–9 preview
 /// keys; the rest are bulleted. With `highlight_cursor`, the `cursor` row is reverse-styled.
 /// Windows to `visible_rows` rows starting at `offset`.
-fn file_rows(
+pub(super) fn file_rows(
     files: &[String],
     cursor: usize,
     offset: usize,
@@ -490,171 +490,6 @@ fn render_compact_gist_bg_vm(
     );
 }
 
-/// The gist detail header from a view model: info line + `Files │ Comments` tabs.
-fn render_detail_header_vm(
-    frame: &mut Frame,
-    area: Rect,
-    detail: &super::view_model::GistDetailVm,
-    chrome: &super::view_model::ChromeVm,
-    theme: &Theme,
-    layout: &mut MouseLayout,
-) {
-    let lines = vec![
-        Line::from(detail.info_line.clone()),
-        detail_focus_tabs_line(detail.focus, theme),
-    ];
-    frame.render_widget(
-        Paragraph::new(lines).style(theme.base_style()).block(
-            Block::default()
-                .title(detail.block_title.clone())
-                .borders(Borders::ALL)
-                .border_type(BorderType::Rounded)
-                .border_style(Style::default().fg(theme.accent))
-                .style(theme.base_style())
-                .padding(Padding::horizontal(1)),
-        ),
-        area,
-    );
-    if chrome.mouse_enabled {
-        // Tab line is the 2nd content row (border + gist-info line above it); content starts
-        // after the left border (1) + horizontal padding (1). Labels: " Files " (7), " │ " (3),
-        // " Comments " (10) — see detail_focus_tabs_line.
-        let content_x = area.x + 2;
-        let tabs_y = area.y + 2;
-        layout.detail_tab_files = Some(Rect::new(content_x, tabs_y, 7, 1));
-        layout.detail_tab_comments = Some(Rect::new(content_x + 10, tabs_y, 10, 1));
-    }
-}
-
-/// Files tab from the view model (full file list; paint windows to the area height).
-fn render_gist_file_list_vm(
-    frame: &mut Frame,
-    area: Rect,
-    detail: &super::view_model::GistDetailVm,
-    chrome: &super::view_model::ChromeVm,
-    theme: &Theme,
-    layout: &mut MouseLayout,
-) {
-    let files = &detail.files;
-    let cursor = detail.file_cursor.min(files.len().saturating_sub(1));
-    let visible_rows = (area.height as usize).saturating_sub(2);
-    let offset = file_list_scroll(cursor, visible_rows, files.len());
-    if chrome.mouse_enabled {
-        layout.detail_files = Some(PaneHit { rect: area, offset });
-    }
-    let lines = file_rows(files, cursor, offset, visible_rows, true, theme);
-    frame.render_widget(
-        Paragraph::new(lines).style(theme.base_style()).block(
-            Block::default()
-                .title(format!("Files ({})", files.len()))
-                .borders(Borders::ALL)
-                .border_type(BorderType::Rounded)
-                .border_style(Style::default().fg(theme.accent))
-                .style(theme.base_style())
-                .padding(Padding::horizontal(1)),
-        ),
-        area,
-    );
-}
-
-/// Comments pane from the view model: styles plain presentation facts and fills hit/scroll layout.
-fn render_gist_comments_vm(
-    frame: &mut Frame,
-    area: Rect,
-    comments: &super::view_model::CommentsPaneVm,
-    theme: &Theme,
-    layout: &mut MouseLayout,
-) {
-    use super::view_model::{CommentLineVm, CommentsAffordance, CommentsPaneVm};
-
-    let mut body: Vec<Line> = Vec::new();
-    let mut affordance_present = false;
-    let mut title = "Comments".to_string();
-    let mut scroll = 0u16;
-
-    match comments {
-        CommentsPaneVm::Loading => body.push(Line::from(Span::styled(
-            "Loading comments…",
-            Style::default().fg(theme.dim),
-        ))),
-        CommentsPaneVm::PromptLoad => body.push(Line::from(Span::styled(
-            "Tab here to load comments",
-            Style::default().fg(theme.dim),
-        ))),
-        CommentsPaneVm::Error { message } => body.push(Line::from(Span::styled(
-            message.clone(),
-            Style::default().fg(theme.del_color),
-        ))),
-        CommentsPaneVm::Empty => body.push(Line::from(Span::styled(
-            "No comments",
-            Style::default().fg(theme.dim),
-        ))),
-        CommentsPaneVm::Thread {
-            title: t,
-            affordance,
-            lines,
-            scroll: s,
-        } => {
-            title = t.clone();
-            scroll = *s;
-            let label = match affordance {
-                CommentsAffordance::LoadingMore => "Loading…",
-                CommentsAffordance::LoadOlder => {
-                    affordance_present = true;
-                    "↑ Load 30 older comments"
-                }
-                CommentsAffordance::StartOfThread => "— Start of thread —",
-            };
-            body.push(Line::from(Span::styled(
-                label,
-                Style::default().fg(theme.dim),
-            )));
-            body.push(Line::from(""));
-            for line in lines {
-                match line {
-                    CommentLineVm::Author { text } => body.push(Line::from(Span::styled(
-                        text.clone(),
-                        Style::default().fg(theme.accent),
-                    ))),
-                    CommentLineVm::Body { text } => body.push(Line::from(text.clone())),
-                    CommentLineVm::Blank => body.push(Line::from("")),
-                }
-            }
-        }
-    }
-
-    layout.comments_load_older = if affordance_present {
-        Some(Rect::new(
-            area.x + 1,
-            area.y + 1,
-            area.width.saturating_sub(2),
-            1,
-        ))
-    } else {
-        None
-    };
-
-    let total_lines = body.len();
-    let inner_rows = area.height.saturating_sub(2);
-    layout.comments_max_scroll = Some((total_lines as u16).saturating_sub(inner_rows));
-
-    frame.render_widget(
-        Paragraph::new(body)
-            .style(theme.base_style())
-            .scroll((scroll, 0))
-            .wrap(Wrap { trim: false })
-            .block(
-                Block::default()
-                    .title(title)
-                    .borders(Borders::ALL)
-                    .border_type(BorderType::Rounded)
-                    .style(theme.base_style()),
-            ),
-        area,
-    );
-    render_text_scrollbar(frame, area, total_lines, scroll as usize);
-}
-
 /// The default (idle) footer hint on screens that used to show a long per-screen key dump.
 pub(super) const MINIMAL_HINT: &str = "; Menu · Ctrl+p Palette";
 
@@ -665,111 +500,6 @@ pub(super) fn footer_with_status(status: Option<&str>, hints: &str) -> (String, 
     match status {
         Some(message) => (message.to_string(), false),
         None => (hints.to_string(), true),
-    }
-}
-
-/// The Files|Comments tab index, mirroring `detail_focus`. Pure so the tab selection is
-/// trivially testable and stays in sync with the navigation handler. Files is the default
-/// tab, so it comes first.
-pub(super) fn detail_focus_tab(focus: DetailFocus) -> usize {
-    match focus {
-        DetailFocus::Files => 0,
-        DetailFocus::Comments => 1,
-    }
-}
-
-/// A `Files │ Comments` focus indicator line, with the pane Tab currently drives highlighted.
-/// Rendered just under the gist's basic info (inside the info box) rather than as a floating
-/// strip, so the active focus is visible without a disconnected top row.
-pub(super) fn detail_focus_tabs_line(focus: DetailFocus, theme: &Theme) -> Line<'static> {
-    let active = detail_focus_tab(focus);
-    let active_style = Style::default()
-        .fg(theme.fg_on_accent)
-        .bg(theme.accent)
-        .add_modifier(Modifier::BOLD);
-    let idle_style = Style::default().fg(theme.dim);
-    let mut spans = Vec::new();
-    for (i, label) in ["Files", "Comments"].iter().enumerate() {
-        if i > 0 {
-            spans.push(Span::styled(" │ ", idle_style));
-        }
-        let style = if i == active {
-            active_style
-        } else {
-            idle_style
-        };
-        spans.push(Span::styled(format!(" {label} "), style));
-    }
-    Line::from(spans)
-}
-
-fn render_gist_detail_vm(
-    frame: &mut Frame,
-    state: &AppState,
-    detail: &super::view_model::GistDetailVm,
-    chrome: &super::view_model::ChromeVm,
-    layout: &mut MouseLayout,
-) {
-    let area = frame.area();
-    let area = render_top_bar(frame, area, &state.theme, chrome.mouse_enabled, layout);
-    // Fixed 4-row header (borders + basic-info line + focus tabs); the active tab — the file
-    // list or the comments, never both — fills the rest above the footer.
-    let chunks = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints([
-            Constraint::Length(4),
-            Constraint::Min(3),
-            Constraint::Length(footer_height(&detail.footer, area.width, "")),
-        ])
-        .split(area);
-    if !detail.missing {
-        render_detail_header_vm(frame, chunks[0], detail, chrome, &state.theme, layout);
-        match detail.focus {
-            DetailFocus::Files => {
-                render_gist_file_list_vm(frame, chunks[1], detail, chrome, &state.theme, layout)
-            }
-            DetailFocus::Comments => {
-                render_gist_comments_vm(frame, chunks[1], &detail.comments, &state.theme, layout)
-            }
-        }
-    }
-    render_footer(
-        frame,
-        chunks[2],
-        "",
-        &detail.footer,
-        detail.footer_colored,
-        &state.theme,
-        layout,
-    );
-
-    let edit_modal = if detail.editing_description {
-        // The modal covers the file list and tabs; drop their hit regions so a click
-        // behind the modal doesn't move the cursor or switch tabs.
-        layout.detail_files = None;
-        layout.detail_tab_files = None;
-        layout.detail_tab_comments = None;
-        layout.comments_load_older = None;
-        Some(render_centered_modal_input(
-            frame,
-            "Edit description (Enter apply · Esc cancel)",
-            "",
-            &state.description_input,
-            "",
-            state.theme.accent,
-            &state.theme,
-        ))
-    } else {
-        None
-    };
-    if chrome.mouse_enabled {
-        // When the edit-description modal is open, the close button belongs on it;
-        // otherwise it sits on the full-screen detail view's top-right corner.
-        layout.close_button = Some(render_close_button(
-            frame,
-            edit_modal.unwrap_or(area),
-            &state.theme,
-        ));
     }
 }
